+++ sdk/python/pulumi_kubernetes/helm/v3/helm.py
@@ -415,6 +415,10 @@ class BaseChartOpts:
     Warning: This option should not be used if you have resources depending on Outputs from the Chart.
     """
 
+    is_upgrade: Optional[pulumi.Input[bool]]
+    """
+    Manually override to tell the provider to run helm template w/ --is-upgrade.
+    """
     def __init__(self,
                  namespace: Optional[pulumi.Input[str]] = None,
                  values: Optional[pulumi.Inputs] = None,
@@ -423,7 +427,8 @@ class BaseChartOpts:
                  api_versions: Optional[Sequence[pulumi.Input[str]]] = None,
                  include_test_hook_resources: Optional[pulumi.Input[bool]] = None,
                  skip_crd_rendering: Optional[pulumi.Input[bool]] = None,
-                 skip_await: Optional[pulumi.Input[bool]] = None):
+                 skip_await: Optional[pulumi.Input[bool]] = None,
+                 is_upgrade: Optional[pulumi.Input[bool]] = None):
         """
         :param Optional[pulumi.Input[str]] namespace: Optional namespace to install chart resources into.
         :param Optional[pulumi.Inputs] values: Optional overrides for chart values.
@@ -442,6 +447,8 @@ class BaseChartOpts:
         :param Optional[pulumi.Input[bool]] skip_await: Skip await logic for all resources in this Chart. Resources
                will be marked ready as soon as they are created. Warning: This option should not be used if you have
                resources depending on Outputs from the Chart.
+        :param Optional[pulumi.Input[bool]] is_upgrade: Manually override to tell the provider
+                to run helm template w/ --is-upgrade.
         """
         self.namespace = namespace
         self.include_test_hook_resources = include_test_hook_resources
@@ -451,6 +458,7 @@ class BaseChartOpts:
         self.transformations = transformations
         self.resource_prefix = resource_prefix
         self.api_versions = api_versions
+        self.is_upgrade = is_upgrade
 
     def to_json(self):
         return pulumi.Output.from_input(self.__dict__).apply(
@@ -485,6 +493,10 @@ class ChartOpts(BaseChartOpts):
     Additional options to customize the fetching of the Helm chart.
     """
 
+    is_upgrade: Optional[pulumu.Input[bool]]
+    """
+    Manually override to tell the provider to run helm template w/ --is-upgrade.
+    """
     def __init__(self,
                  chart: pulumi.Input[str],
                  namespace: Optional[pulumi.Input[str]] = None,
@@ -497,7 +509,8 @@ class ChartOpts(BaseChartOpts):
                  api_versions: Optional[Sequence[pulumi.Input[str]]] = None,
                  include_test_hook_resources: Optional[pulumi.Input[bool]] = None,
                  skip_crd_rendering: Optional[pulumi.Input[bool]] = None,
-                 skip_await: Optional[pulumi.Input[bool]] = None):
+                 skip_await: Optional[pulumi.Input[bool]] = None,
+                 is_upgrade: Optional[pulumi.Input[bool]] = None):
         """
         :param pulumi.Input[str] chart: The name of the chart to deploy.  If `repo` is provided, this chart name
                will be prefixed by the repo name.
@@ -526,9 +539,11 @@ class ChartOpts(BaseChartOpts):
         :param Optional[pulumi.Input[bool]] skip_await: Skip await logic for all resources in this Chart. Resources
                will be marked ready as soon as they are created. Warning: This option should not be used if you have
                resources depending on Outputs from the Chart.
+        :param Optional[pulumi.Input[bool]] is_upgrade: Manually override to tell the provider
+                to run helm template w/ --is-upgrade.
         """
         super(ChartOpts, self).__init__(namespace, values, transformations, resource_prefix, api_versions,
-                                        include_test_hook_resources, skip_crd_rendering, skip_await)
+                                        include_test_hook_resources, skip_crd_rendering, skip_await, is_upgrade)
         self.chart = chart
         self.repo = repo
         self.version = version
@@ -545,6 +560,11 @@ class LocalChartOpts(BaseChartOpts):
     The path to the chart directory which contains the `Chart.yaml` file.
     """
 
+    is_upgrade: Optional[pulumu.Input[bool]]
+    """
+    Manually override to tell the provider to run helm template w/ --is-upgrade.
+    """
+
     def __init__(self,
                  path: pulumi.Input[str],
                  namespace: Optional[pulumi.Input[str]] = None,
@@ -554,7 +574,8 @@ class LocalChartOpts(BaseChartOpts):
                  api_versions: Optional[Sequence[pulumi.Input[str]]] = None,
                  include_test_hook_resources: Optional[pulumi.Input[bool]] = None,
                  skip_crd_rendering: Optional[pulumi.Input[bool]] = None,
-                 skip_await: Optional[pulumi.Input[bool]] = None):
+                 skip_await: Optional[pulumi.Input[bool]] = None,
+                 is_upgrade: Optional[pulumi.Input[bool]] = None):
         """
         :param pulumi.Input[str] path: The path to the chart directory which contains the
                `Chart.yaml` file.
@@ -575,10 +596,12 @@ class LocalChartOpts(BaseChartOpts):
         :param Optional[pulumi.Input[bool]] skip_await: Skip await logic for all resources in this Chart. Resources
                will be marked ready as soon as they are created. Warning: This option should not be used if you have
                resources depending on Outputs from the Chart.
+        :param Optional[pulumi.Input[bool]] is_upgrade: Manually override to tell the provider
+                to run helm template w/ --is-upgrade.
         """
 
         super(LocalChartOpts, self).__init__(namespace, values, transformations, resource_prefix, api_versions,
-                                             include_test_hook_resources, skip_crd_rendering, skip_await)
+                                             include_test_hook_resources, skip_crd_rendering, skip_await, is_upgrade)
         self.path = path