+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `PULUMI_K8S_ENABLE_PATCH_FORCE` env var support (https://github.com/pulumi/pulumi-kubernetes/pulls/2260)
 - Add link to resolution guide for SSA conflicts (https://github.com/pulumi/pulumi-kubernetes/pulls/2265)
+- Add isUpgrade flag to helm chart options. (https://github.com/pulumi/pulumi-kubernetes/pull/2276)
 
 ## 3.23.0 (December 8, 2022)